Adding timeout for CAM objects and solving station id bug

+++ cam_collision_detection/include/cam_collision_detection/cam_collision_detection.hpp
@@ -34,7 +34,7 @@ public:
   
   bool isCVInRange(PredictedObject ego, PredictedObject cv);
 
-  uint16_t getStationID(unique_identifier_msgs::msg::UUID object_id);
+  uint32_t getStationID(unique_identifier_msgs::msg::UUID object_id);
 
 private:
   /* ROS2 entities */
